Qol changes
After choosing an action the robot will do that action for a very short time to make it easier to train the robot. It will also no longer try to scan again after a scan has been saved because this didn't work. The time has NOT been tested but will be soon to make sure it is reasonable