Create AITrainingV3
NOT TESTED.
A way to collect training data without having to re-run the script each time, should be a lot faster. The others usually have to be started several times due to errors from the lidar, but this only needs to be started once and can then collect a lot of scans if it works.